+++ src/schemas/images.py
@@ -1,6 +1,9 @@
 from datetime import datetime
 from typing import List, Optional
 from pydantic import BaseModel, Field
+from src.schemas.comments import CommentResponse
+from src.schemas.tags import TagResponse
+
 
 class UserForImage(BaseModel):
     id: int
@@ -10,26 +13,6 @@ class UserForImage(BaseModel):
     class Config:
         from_attributes = True
 
-class CommentBase(BaseModel):
-    comment: str
-
-class CommentResponse(CommentBase):
-    id: int
-    user: UserForImage
-
-    class Config:
-        from_attributes = True
-
-
-class TagModel(BaseModel):
-    name: str = Field(max_length=25)
-
-
-class TagResponse(TagModel):
-    id: int
-
-    class Config:
-        from_attributes = True
 
 class ImageBase(BaseModel):
     image: str = Field(max_length=255)
